Remove GitLab CI configuration and update project settings
The GitLab CI configuration file (.gitlab-ci.yml) was removed as it's not required anymore. It was replaced by various project configuration files in the .idea directory that were added. These included run configurations, module settings, JSON schema configurations and others which are deemed necessary for the current requirements of the project.